Advanced Coin Flip
In this game, there are 2 modes: Standard and Competitive. Standard is a basic coin flip. When you enter N/n it gives you a random heads or tails, but competitive is basically making this game single player against the computer. The objective is to guess what the computer rolls, so if you guess correctly, you get a congratulations display.